alt text boolean

+++ R/add_tables.R
@@ -130,20 +130,24 @@ add_tables <- function(
     )
   }
 
-  intermediate_tabs_docx <- gsub(".docx", "-inttabs.docx", docx_out)
+  if (isTRUE(config$add_alt_text)) {
+    intermediate_tabs_docx <- gsub(".docx", "-inttabs.docx", docx_out)
 
-  print(document, target = intermediate_tabs_docx)
+    print(document, target = intermediate_tabs_docx)
 
-  add_tables_alt_text(
-    intermediate_tabs_docx,
-    docx_out
-  )
+    add_tables_alt_text(
+      intermediate_tabs_docx,
+      docx_out
+    )
+
+    unlink(intermediate_tabs_docx)
+    log4r::debug(.le$logger, "Deleting intermediate tabs document")
+  } else {
+    print(document, target = docx_out)
+  }
 
   unlink(intermediate_docx)
   log4r::debug(.le$logger, "Deleting intermediate document")
-
-  unlink(intermediate_tabs_docx)
-  log4r::debug(.le$logger, "Deleting intermediate tabs document")
 
   log4r::info(.le$logger, paste0("Final document saved to: ", docx_out))
